+++ NEWS.md
@@ -2,6 +2,8 @@ Version 0.11.2 For Dev-C++ 7 Beta
  - fix: button "run all problem cases" not disabled when compiling or debugging
  - enhancement: set font for problem case input/output textedits
  - enhancement: when run problem cases, updates output immediately
+ - fix: current line of the disassembly in the cpu window not correctly setted
+ - enhancement: add "step into one machine instruction" and "step over one machine instruction" in the cpu window
 
 Version 0.11.1 For Dev-C++ 7 Beta
  - enhancement: Problem's test case shouldn't accept rich text inputs